Move Planetfall companion-chat classes out of shared ChatLambda (#485)
ChatWithFloyd, ChatWithBlather, ChatWithAmbassador, and IChatWithFloyd are
Planetfall-specific characters but lived in the shared ChatLambda project
(referenced by GameEngine), leaking game-specific code into the shared layer.

Move all four into the Planetfall project under a new Planetfall.AI namespace.
The generic base (ChatWithCompanion) and its CompanionResponse/CompanionMetadata
records stay in ChatLambda, so consumers keep `using ChatLambda;` and add
`using Planetfall.AI;`.

Pure refactor — no behavior change:
- Class bodies, method names, and the AssistantName constants are untouched.
- The new AI/ folder is excluded from Planetfall's hint-knowledge embed
  (**/*.cs), keeping that bundle byte-identical (these files were never
  embedded when they lived in ChatLambda).

All UnitTests (1083) and Planetfall.Tests (3207) pass; full solution builds
with 0 errors.
