+++ Planetfall.Tests/Walkthrough/WalkthroughTestBase.cs
@@ -7,6 +7,7 @@ using GameEngine;
 using JetBrains.Annotations;
 using Model.Interface;
 using Moq;
+using Planetfall.AI;
 using Planetfall.Item.Feinstein;
 using Planetfall.Item.Kalamontee.Mech;
 using Planetfall.Item.Kalamontee.Mech.FloydPart;
