+++ Planetfall/AI/ChatWithBlather.cs
@@ -1,6 +1,7 @@
 using Amazon.Lambda;
+using ChatLambda;
 
-namespace ChatLambda;
+namespace Planetfall.AI;
 
 public class ChatWithBlather(IAmazonLambda? lambdaClient) : ChatWithCompanion(lambdaClient)
 {
